feat: updated perps deposit in progress to fire on perpsDeposit approval
